fixed a serious problem with function call (run_function).
We pushed the current symbol table before function call, evaluated every parameter into the current symbol table, and then called the function. If the parameters used variables in the previous symbol table, this would fail. Now we push (because we need it to be pushed to point by reference), but keep the variables. define function variables, replace current symbol table with function symbol table, call, and then pop the old one.

this is not correct still if any changes to the variables is done in function params (side effect when calling)