added call_function public function in emulator, used to call any function (internal, mocked, emulated).
also refactored run_sub to accept  as array of values instead of ParserNode as well, so that normal function calls (specially for calling emulated functions with direct values) is supported.

this was all needed to mock callback functions such as array_map, that need to call that callback internally. array_map works like a charm now, but is sort of complicated