fixed a very important and delicate problemgit add -u the current_* variables of the emulator were updated when calling a function/method, but prior to evaluating the arguments of that function. This meant that those arguments would be evaluated in the wrong context, e.g if self, this or current_function was used on those arguments, it would match the new object or context instead of the old one.
to handle it, all of these should be set by run_function function, after the function prologue and prior to executing the function code, and should be restored after the execution of the function code.
this resulted in run_function receiving two arrays now, one is wrappings, which is a key/value array that will be set as current_{}= for the emulator, and another trace_args which will update backtrace if needed.